Fix issues when starting the app without giving path
The database was created right at the beginning of the main where the
user selected directory was not available if it wasn't given as a
command line argument.

Now the database is created after the directory has been selected and
the internal DB collections are created lazily.

Also fixed some related issues with creating directories and loading
categories.