Optimize loading images at startup
Don't reload metadata for images if the image hasn't changed. Also
changed the logic how meta data is read in. Now the meta data is removed
for each image instead of truncating the whole meta data table. This
makes it possible to optimize the meta data loading as the up-to-date
meta data is not removed.