Refactor zooming
Now it supports zoom levels that are not in the select box. The zoom
level step calculation still needs some work. The zoom should not be
fully linear and it should round to nicer values. Now it's possible to
have zoom levels like 99% or 101% when 100% would be more appropriate.

Also zooming in higher zoom levels is too slow so the steps should be
larger.