Remove directory from DB
From now on the DB won't store the directory anymore. Having the
directory in the DB makes moving the image directory harder if the DB
has already been initialized. There should not be a need to store the
information in the database. We can just assume that the work dir is
relative to the folder's database. Or actually the DB is relative to the
work dir and that work dir is the base path for all operations.